Rename elo<...>
