+++ docs/modules/Rand.html
@@ -15,7 +15,7 @@
     <div class="uk-container uk-container-center">
     <div class="uk-grid uk-margin-top">
 
-    <img src="../red_putit.png" alt="elomod" class="icon">
+    <img src="../red_putit.png" alt="Lua API" class="icon">
 
     <!-- sidebar start -->
     <aside class="uk-width-medium-1-5 uk-row-first">
@@ -27,7 +27,7 @@
 
     <!-- project title -->
 
-    <a href="../index.html"><h1 style="padding-top: 8px">elomod</h1></a>
+    <a href="../index.html"><h1 style="padding-top: 8px">Lua API</h1></a>
 
     <!-- project description -->
     <p class="project-desc">ElonaFoobar Mod API</p>
