+++ tools/i18n_checker/src/hclua/util.lua
