+++ tools/i18n_checker/src/inspect.lua
